int -> size_t

+++ Lab1CPP/task1.cpp
@@ -11,7 +11,7 @@ void task1(Array *arr)
     double average = sum / double(arr->size());
     double dif(sum);
     int closest = 0;
-    for (int i = 0; i < arr->size(); i++) {
+    for (size_t i = 0; i < arr->size(); i++) {
         double tmp = std::abs(average - arr->get(i));
         if (dif > tmp) {
             dif = tmp;
